+++ package.json
@@ -52,7 +52,7 @@
     "@types/jest": "^24.0.23",
     "@types/jsdom": "^12.2.4",
     "@types/loadable__component": "^5.10.0",
-    "@types/lodash": "^4.14.149",
+    "@types/lodash.chunk": "^4.2.6",
     "@types/oboe": "^2.0.28",
     "@types/qs": "^6.5.3",
     "@types/react-redux": "^7.1.1",
@@ -96,7 +96,7 @@
     "jest-extended": "^0.11.2",
     "jest-fetch-mock": "^2.1.2",
     "jsdom": "^11.12.0",
-    "lodash": "^4.17.15",
+    "lodash.chunk": "^4.2.0",
     "mini-css-extract-plugin": "^0.8.0",
     "prettier": "^1.19.1",
     "raf": "^3.4.1",
